+++ nix/neovim/config/after/plugin/cmp.lua
@@ -1,0 +1,10 @@
+local cmp = require("cmp")
+
+cmp.setup({
+	mapping = cmp.mapping.preset.insert({
+		["<C-Space>"] = cmp.mapping.complete(),
+	}),
+	sources = cmp.config.sources({
+		{ name = "nvim_lsp" },
+	}),
+})