+++ nix/neovim/config/after/plugin/marks.lua
@@ -1,0 +1,1 @@
+vim.cmd.highlight("SignatureMarkText guifg=#816DBD ctermfg=4")