+++ nix/neovim/config/init.lua
@@ -1,0 +1,12 @@
+-- Immediate combo changes
+require("user.remap")
+
+-- Preferences for editing
+require("user.set")
+require("user.style")
+
+-- Customizations for files
+require("user.files")
+
+-- Greeetings
+print("hello!")