+++ nix/neovim/config/lua/user/files/c.lua
@@ -1,0 +1,9 @@
+vim.api.nvim_create_autocmd({
+	"BufRead",
+	"BufNewFile",
+}, {
+	pattern = "*.cl",
+	callback = function()
+		vim.cmd.set("filetype=c")
+	end,
+})