+++ nix/neovim/config/lua/user/files/init.lua
@@ -1,0 +1,3 @@
+require("user.files.c")
+require("user.files.ecmascript")
+require("user.files.groovy")