+++ nvim/after/plugin/git.lua
@@ -1,3 +1,0 @@
-vim.cmd.highlight "GitGutterAdd guifg=#2f7355 ctermfg=2"
-vim.cmd.highlight "GitGutterChange guifg=#4d798f ctermfg=3"
-vim.cmd.highlight "GitGutterDelete guifg=#b84848 ctermfg=1"