+++ nvim/after/plugin/marks.lua
@@ -1,1 +1,0 @@
-vim.cmd.highlight "SignatureMarkText guifg=#816DBD ctermfg=4"