+++ nvim/lua/configs/init.lua
@@ -1,4 +1,0 @@
-require("configs.packer")
-require("configs.remap")
-require("configs.set")
-require("configs.style")